feat: update oath-toolkit to 2.6.9

+++ pam_oath/tests/run-root-tests.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# Copyright (C) 2011-2021 Simon Josefsson
+# Copyright (C) 2011-2023 Simon Josefsson
 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -22,6 +22,17 @@ fi
 
 srcdir=${srcdir:-.}
 
+FAKETIME=datefudge
+TSTAMP=`TZ=UTC $FAKETIME "2006-09-23" date -u +%s`
+if test "$TSTAMP" != "1158969600"; then
+    FAKETIME=faketime
+    TSTAMP=`TZ=UTC $FAKETIME "2006-09-23" date -u +%s`
+    if test "$TSTAMP" != "1158969600"; then
+	echo "Faketime or datefudge missing ($TSTAMP)" >&2
+	exit 77
+    fi
+fi
+
 ETCPAMCFG=/etc/pam.d/pam_oath1
 ETCUSRCFG=/etc/tst-pam_oath.users
 
@@ -37,7 +48,18 @@ if test -f $ETCUSRCFG; then
     exit 1
 fi
 
-echo "auth requisite pam_oath.so debug usersfile=$ETCUSRCFG window=20 digits=6" > $ETCPAMCFG
+so_path_rel="${srcdir}/../.libs/pam_oath.so"
+so_path="$(readlink -f "${so_path_rel}")"
+if test -z "${so_path}"; then
+    echo "Unable to resolve path to pam_oath.so: ${so_path_rel}"
+    exit 1
+fi
+if ! test -f "${so_path}"; then
+    echo "pam_oath.so not found at: ${so_path}"
+    exit 1
+fi
+
+echo "auth requisite [${so_path}] debug usersfile=$ETCUSRCFG window=20 digits=6" > $ETCPAMCFG
 echo "HOTP user1 - 00" > $ETCUSRCFG
 echo "HOTP user2 pw 00" >> $ETCUSRCFG
 echo "HOTP/T30 user3 - 00" >> $ETCUSRCFG
@@ -52,18 +74,11 @@ if ! test -f $ETCUSRCFG; then
     exit 77
 fi
 
-TSTAMP=`datefudge "2006-09-23" date -u +%s`
-if test "$TSTAMP" != "1158962400"; then
-    echo "Cannot fake timestamp, install datefudge to check better. ($TSTAMP)"
-    ./test-pam_oath-root user3
-    rc=$?
-else
-    datefudge 2006-12-07 ./test-pam_oath-root user3
-    rc=$?
-fi
+TZ=UTC $FAKETIME "2006-12-07" ./test-pam_oath-root user3
+rc=$?
 
 if test "$rc" != "77"; then
-    diff -u $srcdir/expect.oath $ETCUSRCFG || rc=1
+    diff -u "$srcdir/expect.oath" $ETCUSRCFG || rc=1
 fi
 
 rm -f $ETCPAMCFG $ETCUSRCFG
